agents/rx_agent.py
from strands import Agent
from strands_tools import http_request
from agents.config import bedrock_model

rx_agent = Agent(
    model=bedrock_model,
    system_prompt="""You are a prescription drug lookup assistant. When given a drug name, you retrieve RxNorm and FDA data using the http_request tool.

For RxNorm data, call: https://rxnav.nlm.nih.gov/REST/rxcui.json?name={term}
For FDA data, call: https://api.fda.gov/drug/label.json?search=openfda.generic_name:{term}&limit=1

Be direct and include source data using quotation formatting:
- RxCUI: "123456"
- Brand Name: "Advil"
- Purpose: "Pain relief"
- Warnings: "Do not exceed recommended dose"

Only return fields that are found. Format the response in a structured way that's suitable for front-end display with highlighting or bolding.""",
    tools=[http_request]
) 